refactor(storage): narrow SPARQL deadline boundary
